qemu: Use writeback cache
With `cache=none`, we were seeing errors like this:

    qemu-system-x86_64: -drive file=/home/dxu/dev/vmtest/tests/.assets/image-ubuntu22.raw,format=raw,index=1,media=disk,if=virtio,cache=none:
    Cannot get 'write' permission without 'resize': Image size is not a multiple of request alignment

There are some hints here about the error:

    https://patchwork.kernel.org/project/qemu-devel/patch/20210819101200.64235-1-hreitz@redhat.com/#24392517

Basically it looks like for O_DIRECT to work we need an aligned image
size. That seems like an odd restriction, so let's just use a writeback
cache instead. That is the default setting anyways so it should be fine.