qga: Retry the connection every second.
If we connect to the QGA socket too early, the VM's QGA will never
respond to us no matter how many times we write commands to it.

So, do the guest_sync first synchronously and recreate the socket each
attempt until it passes. Then hand off to the thread loop.