run kernel in chroot
Insteads of mounting the local FS'root, we can do the same but in a chroot.
The main benefit would be to be able to run a kernel/artifacts in a guest FS
which may be different then the host FS. For instance when collecting CI artifacts
and trying to reproduce locally.
CI may run in say Ubuntu 22.04 , while my host may be Fedora 38.
This also opens up for a next diff where we can run againdt a foreign architecture
provided that the chrrot is for that arch.
```
$ cat /etc/debian_version && RUST_LOG=debug cargo run -- -k $KERNEL_REPO/arch/x86_64/boot/bzImage -r debian 'cat /etc/debian_version'
trixie/sid
   Compiling vmtest v0.8.3 (/home/chantra/devel/danobi-vmtest)
    Finished dev [unoptimized + debuginfo] target(s) in 8.11s
     Running `target/debug/vmtest -k /home/chantra/devel/bpf-next//arch/x86_64/boot/bzImage -r debian 'cat /etc/debian_version'`
=> bzImage
===> Booting
===> Setting up VM
===> Running command
11.5
bash: line 1: cd: /home/chantra/devel/danobi-vmtest: No such file or directory
```