ci: Set kvm perms
Been getting a lot of the following error in CI runs lately:

    [tests/test.rs:91] get_error(recv, None) = Some(
        Error {
            context: "Could not access KVM kernel module: Permission denied\nqemu-system-x86_64: failed to initialize kvm: Permission denied\n",
            source: Error {
                context: "Failed to connect QMP",
                source: Os {
                    code: 111,
                    kind: ConnectionRefused,
                    message: "Connection refused",
                },
            },
        },
    )

It happens on and off for no good reason. No idea why. I'm guessing GHA
might be slowly rolling out nested virt to non-large runners?

The fix in this commit is required for large runners to use nested virt.
Maybe we need this for regular runners now.

See https://github.blog/changelog/2023-02-23-hardware-accelerated-android-virtualization-on-actions-windows-and-linux-larger-hosted-runners/