qemu: Send guest command output to virtio-serial port
This teaches the command entrypoint to locate the virtio-serial port and
stream all output (with stdout and stderr merged) to the chardev. This
gives us streaming command output which is a much nicer UX for the user.

But b/c we support virtual machine images (as well as plain kernel), we
make the virtio-serial port optional. B/c we don't require folks build
VM image kernels with CONFIG_VIRTIO_CONSOLE.