Return RC code of command when running one-liner
If we don't even get to run the command because of some other issues, EX_UNAVAILABLE
(which is kind of the catchall for "something happened") is returned instead.

When multiple targets are ran, we can't really know what to return.... in which
case the old behaviour is kept, e.g return the number of failed runs.

Added test to validate the behaviour.

Fixes #43

Signed-off-by: Manu Bretelle <chantr4@gmail.com>