[ui][run_one] Return a Result<i32> instead of TargetRcUi
If an error happen when trying to set up/run the VM, return an Error,
otherwise return the return code of the command.

Do not report a non-zero return code from the command as an Error, the command
actually ran, but do log the error to the console to make it more obvious to
the user.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>