[qemu] Multiplex serial and monitor over hostr stdio
By using `-serial mon:stdio` we multiplex the guest'serial console and QEMU
monitor to the host's stdio. This gives us 2 features in interactive mode:

1) Access to the monitor via the `Ctrl-c a` keybinding
2) Intercept Ctrl-c and pass it to the guest instead of killing the guest.

In non-interactive mode, stdio is not hooked to qemu and this is a no-op.

https://www.qemu.org/docs/master/system/qemu-manpage.html#hxtool-9

```
mon:dev_string
This is a special option to allow the monitor to be multiplexed onto another serial port. The monitor is accessed with key sequence of Control-a and then pressing c. dev_string should be any one of the serial devices specified above. An example to multiplex the monitor onto a telnet server listening on port 4444 would be:

-serial mon:telnet::4444,server=on,wait=off

When the monitor is multiplexed to stdio in this way, Ctrl+C will not terminate QEMU any more but will be passed to the guest instead.
```

Testing:

In interactive mode, typed `Ctrl-a h`
```
root@(none):/#
C-a h    print this help
C-a x    exit emulator
C-a s    save disk data back to file (if -snapshot)
C-a t    toggle console timestamps
C-a b    send break (magic sysrq)
C-a c    switch between console and monitor
C-a C-a  sends C-a
```
and confirmed that Ctrl-c did not kill the guest.

Signed-off-by: Manu Bretelle <chantr4@gmail.com>