Always capture stderr
Output of the qemu process on stderr is only captured when running in
non-interactive mode. However, the error reporting infrastructure always
assumes that it is being captured. As a result, we may see a panic
instead of a proper error report when, for one reason or another, Qemu
fails to start (as an example).
Fix the issue by making sure to always capture stderr. I haven't found
any adverse effects of doing so in my admittedly limited testing.

Closes: #70

Signed-off-by: Daniel Müller <deso@posteo.net>