Eagerly close temporary init script file
Various of my qemu instances have hung and I had to kill them
forcefully. As a result, I ended up with 10+ vmtest-initGDllO.sh files
in my /tmp/ folder.
I think we don't have to rely on destructor based clean up in this
instance: after the end of the setup phase the init script should no
longer be accessed via file system path and so we can just unlink it.
This way, no matter how the process dies after that point, the temporary
file will always be gone as well.

Signed-off-by: Daniel Müller <deso@posteo.net>