Remove a bunch more clones
We the Qemu::new() constructor might as well just consume the Target
object, given that it won't be used afterwards. Adjust the signature and
then just pull out the members we need instead of cloning them, removing
a bunch of unnecessary allocations.
Similarly, the CommandContext that we instantiate just for the sake of
conveying some data to the template does not need to contain any owned
objects. So just work with references to existing ones, eliminating a
bunch of more clones.

Signed-off-by: Daniel Müller <deso@posteo.net>