init: Set QEMU state dir/PID file path directly
This fixes https://github.com/danobi/vmtest/issues/101 for me.

I'm not sure why I didn't used to hit this, but on my host system
/var/local/run is owned by a privilged user and has no global write
perms, but is also the default where the qemu-ga build writes these
files to.

Fix is just to set it explicitly to the default that exists on other
systems, which the init script is already set up to handle.