qga: Use 80s timeout even for KVM
With all the debug options turned on in the kernel, it can take ~60s
even on my relatively powerful machine (Ryzen 5950x). 80s ought to be
enough for now.

Recently we landed a change to make the error quicker when the kernel
panics, so this change in timeout shouldn't affect much.