ui: Mark failed targets as failed for config interface
Commit 7b29d3a ("[ui][run_one] Return a Result<i32> instead of
TargetRcUi") introduced a bug where failed targets under the config
interface were being marked as PASS. This is confusing - all evidence of
failure is hidden from the terminal.

The previous logic was also pretty confusing b/c it used several state
variables with multiple (and invalid) combinations. Simplify all of it
to use an enum so all the states are more obvious.

This closes #103.