ui: Use Option<i32> over Result<i32>
The result was being used like an option and not like a real result. We
don't need a real result in the UI b/c the UI prints errors to screen
directly. So things like attaching error context is completely
unnecessary.

Switch to Option<T> to make intention more obvious.