ui: Do not strip styling for unattended output
The original need for stripping ansi escape sequences was so we could
apply our own styling when the terminal is attended and we're doing
custom terminal drawing. However, when the output is unattended, we
do not have a need for stripping ansi escape sequences.

Lots of CIs force coloring even though output is not attended. This is
so the logs are easier to read. vmtest should respect that and not strip
ansi escape sequences when output is unattended.